+++ .php-cs-fixer.dist.php
@@ -40,8 +40,8 @@ $rules = [
     'strict_param' => true,
     'php_unit_strict' => true,
     'no_useless_else' => true,
-    'no_useless_return' => true,
     'no_unused_imports' => true,
+    'no_useless_return' => true,
     'strict_comparison' => true,
     'ordered_interfaces' => true,
     'final_internal_class' => true,
@@ -50,13 +50,17 @@ $rules = [
     'list_syntax' => ['syntax' => 'short'],
     'array_syntax' => ['syntax' => 'short'],
     'ordered_class_elements' => ['sort_algorithm' => 'alpha'],
-    'phpdoc_to_comment' => ['allow_before_return_statement' => true],
+    'attribute_empty_parentheses' => ['use_parentheses' => false],
+    'native_function_invocation' => ['include' => ['@internal', '@all', '@compiler_optimized']],
     'php_unit_test_case_static_method_calls' => ['call_type' => 'self'],
-    'native_function_invocation' => ['include' => ['@internal', 'all']],
     'new_with_braces' => ['anonymous_class' => false, 'named_class' => true],
     'ordered_imports' => ['imports_order' => ['const', 'class', 'function']],
     'blank_line_before_statement' => ['statements' => ['declare', 'try', 'return']],
     'header_comment' => ['header' => $comment, 'location' => 'after_open'],
+    'phpdoc_to_comment' => [
+        'allow_before_return_statement' => true,
+        'ignored_tags' => ['phpstan-var', 'phpstan-param', 'phpstan-ignore'],
+    ],
 ];
 
 $paths = [
@@ -69,7 +73,7 @@ $files = [
     __DIR__ . '/rector.php',
 ];
 
-$skippedPaths = [
+$exclude = [
     'font',
     'fonts',
     'Legacy',
@@ -79,7 +83,7 @@ $skippedPaths = [
 $finder = Finder::create()
     ->in($paths)
     ->append($files)
-    ->notPath($skippedPaths);
+    ->exclude($exclude);
 
 $config = new Config();