+++ src/Enums/PdfColorSpace.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This file is part of the 'fpdf' package.
+ *
+ * For the license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @author bibi.nu <bibi@bibi.nu>
+ */
+
+declare(strict_types=1);
+
+namespace fpdf\Enums;
+
+/**
+ * The PDF image color space enumeration.
+ */
+enum PdfColorSpace: string
+{
+    case DEVICE_CMYK = 'DeviceCMYK';
+    case DEVICE_GRAY = 'DeviceGray';
+    case DEVICE_RGB = 'DeviceRGB';
+    case INDEXED = 'Indexed';
+}